Creation of README on "master" branch
Due to the confusion caused by git's local repo naming, I had to push my local repo to Github named as "master" and not "main". As I'm still in very early stages, I'm not able to fix this whilst having projects to finish, so for practical purposes, this project will have a "master" branch instead of "main".